Major rework of permissions, cleaned up some old code.

+++ commands/staff/ban.ts
@@ -2,7 +2,7 @@ import * as commando from 'discord.js-commando'
 import { oneLine } from 'common-tags'
 import * as Logger from '../../util/Logger'
 import * as moment from 'moment'
-import { Message, TextChannel, GuildChannel } from 'discord.js';
+import { Message, TextChannel, GuildChannel, User } from 'discord.js';
 
 module.exports = class BanCommand extends commando.Command {
   constructor(client) {
@@ -20,7 +20,7 @@ module.exports = class BanCommand extends commando.Command {
       guildOnly: true,
 
       args: [{
-        key: 'mention',
+        key: 'user',
         label: 'mention',
         prompt: 'Who would you like to ban?',
         type: 'user',
@@ -29,31 +29,36 @@ module.exports = class BanCommand extends commando.Command {
     })
   }
 
-  async run(msg: commando.CommandMessage, { mention }): Promise<Message | Message[]> {
+  async run(msg: commando.CommandMessage, { user }: { user: User }): Promise<Message | Message[]> {
     if (!msg.member.hasPermission("BAN_MEMBERS")) {
-      await msg.reply("Sorry, but you're a nerd and can't ban members.")
+      await msg.reply("You can't ban users.")
       return msg.delete()
     }
 
-    const memberBanned = msg.guild.member(mention)
-    const userBanned = memberBanned.user
+    const member = msg.guild.member(user)
 
-    try {
-      msg.guild.members.ban(memberBanned)
-      const channel = msg.guild.channels.find((channel: GuildChannel) => channel.name === 'machobot-audit') as TextChannel
+    if (msg.member.roles.highest.comparePositionTo(member.roles.highest) < 0) {
+      await msg.reply('You can\'t ban that user.')
+    }
 
-      if (channel) {
-        channel.send(`${msg.author.username} has banned ${memberBanned} from ${msg.guild.name}.`)
-      }
+    const channel = msg.guild.channels.find((channel: GuildChannel) => channel.name === 'machobot-audit') as TextChannel
+    const banResponse = await msg.guild.members.ban(member).catch(() => {
+      return
+    })
 
-      let time = moment().format('YYYY-MM-DD HH:mm:ss Z')
-      Logger.log(`\r\n[${time}] ${msg.author.username} has banned ${memberBanned} from ${msg.guild.name}.`)
-    } catch (err) {
-      await msg.reply(`Yo ${msg.author} I couldn't ban because of this stupid thing: ${err}`)
+    if (!banResponse) {
+      await msg.reply('I can\'t ban that user. Sorry about that.')
       return msg.delete()
     }
 
-    await msg.reply(userBanned.tag + " has been banned!")
+    if (channel) {
+      channel.send(`${msg.author.username} has banned ${member} from ${msg.guild.name}.`)
+    }
+
+    let time = moment().format('YYYY-MM-DD HH:mm:ss Z')
+    Logger.log(`\r\n[${time}] ${msg.author.username} has banned ${member} from ${msg.guild.name}.`)
+
+    await msg.reply(user.tag + " has been banned!")
     return msg.delete()
   }
 }
